mathlib - handling zero vector when normalizing particle system - fixing small bugs physics - clip function returns if clipPoly is empty, needs more investigation, this should not happen

+++ code/libs/Particle/Particle.h
@@ -9,7 +9,7 @@ public:
 	mwm::Vector3 pos, speed;
 	mwm::Vector4F color;
 	float size, angle, weight;
-	float lifeTime;
+	double lifeTime;
 	double cameraDistance;
 
 	bool operator<(Particle& that){
